+++ database/seeds/DepartmentsTableSeeder.php
@@ -13,6 +13,7 @@ class DepartmentsTableSeeder extends Seeder
     {
         DB::table('departments')->insert([
             'name' => 'none',
+            'status' => 1,
             'created_at' => date("Y-m-d H:i:s"),
             'updated_at' => date("Y-m-d H:i:s")
         ]);