+++ resources/lang/en/main_menu.php
@@ -8,9 +8,27 @@
         'menu_user_change_password' => 'Change password...',
         'menu_user_logout' => 'Logout',
 
+        /**
+         * Orders menu
+         */
+        'menu_orders' => 'Orders',
+        'menu_orders_new' => 'New order',
+        'menu_orders_list' => 'View orders',
+
+        /**
+         * Operator menu
+         */
+        'menu_operator_dictionaries' => 'Dictionaries',
+        'menu_operator_buildings' => 'Buildings',
+        'menu_operator_companies' => 'Companies',
+
+        /**
+         * Menu administrator
+         */
         'menu_administrator' => 'Administrator',
         'menu_administrator_new_user' => 'New User',
         'menu_administrator_users_list' => 'Users list',
+        'menu_administrator_departments' => 'Departments',
 
         /**
          * About application