+++ resources/lang/en/orders_status.php
@@ -1,0 +1,8 @@
+<?php
+    return [
+        'temp' => 'Temporary',
+        'active' => 'Active',
+        'deleted' => 'Deleted',
+        'signed' => 'Signed',
+        'closed' => 'Closed'
+    ];