+++ resources/views/auth/new_order.blade.php
@@ -1,0 +1,10 @@
+@extends('auth.auth_layout')
+
+@section('content')
+    <div class="row">
+        <div class="col">
+            <h3>{{ __('auth.orders_new_order') }}</h3>
+        </div>
+    </div>
+@endsection
+